ENH: Added group analysis.
Changes to accomodate this were made to both analyze_constant (which is now
used as part of the group analysis) and to the tools, to allow conditional
definition of the gui class, depending on the surrounding context (har,
har). That is, whether or not wx is present to be imported. Seems not to have
broken the experiment (toi, toi, toi).