RF: Additional flexibility in the analysis.
Restructured the tools part of the fit to separate different parts of what was
once the mega-function analyze_constant.

This in turn allowed me to more robustly go in and make group_analysis more
flexible, so that it can now accomodate fitting of different functions to the
psychometric curves with different numbers of parameters.